Fix(Renaming the rest of the files and fixing websocket types)
