+++ src/main/java/br/ufrgs/inf/pet/dinoapi/controller/faq/FaqUserQuestionControllerImpl.java
@@ -1,10 +1,10 @@
 package br.ufrgs.inf.pet.dinoapi.controller.faq;
 
 import br.ufrgs.inf.pet.dinoapi.controller.synchronizable.SynchronizableControllerImpl;
-import br.ufrgs.inf.pet.dinoapi.entity.faq.TreatmentQuestion;
-import br.ufrgs.inf.pet.dinoapi.model.faq.FaqUserQuestionDataModel;
+import br.ufrgs.inf.pet.dinoapi.entity.treatment.TreatmentQuestion;
+import br.ufrgs.inf.pet.dinoapi.model.treatment.TreatmentQuestionDataModel;
 import br.ufrgs.inf.pet.dinoapi.repository.faq.FaqUserQuestionRepository;
-import br.ufrgs.inf.pet.dinoapi.service.faq.FaqUserQuestionServiceImpl;
+import br.ufrgs.inf.pet.dinoapi.service.treatment.TreatmentQuestionServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,9 +12,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/faq_user_question/")
 public class FaqUserQuestionControllerImpl extends SynchronizableControllerImpl<
-        TreatmentQuestion, Long, FaqUserQuestionDataModel, FaqUserQuestionRepository, FaqUserQuestionServiceImpl> {
+        TreatmentQuestion, Long, TreatmentQuestionDataModel, FaqUserQuestionRepository, TreatmentQuestionServiceImpl> {
     @Autowired
-    protected FaqUserQuestionControllerImpl(FaqUserQuestionServiceImpl service) {
+    protected FaqUserQuestionControllerImpl(TreatmentQuestionServiceImpl service) {
         super(service);
     }
 }
