+++ src/main/java/br/ufrgs/inf/pet/dinoapi/model/treatment/TreatmentQuestionDataModel.java
@@ -1,4 +1,4 @@
-package br.ufrgs.inf.pet.dinoapi.model.faq;
+package br.ufrgs.inf.pet.dinoapi.model.treatment;
 
 import br.ufrgs.inf.pet.dinoapi.model.synchronizable.SynchronizableDataLocalIdModel;
 
@@ -7,7 +7,7 @@ import javax.validation.constraints.Size;
 
 import static br.ufrgs.inf.pet.dinoapi.constants.FaqConstants.*;
 
-public class FaqUserQuestionDataModel extends SynchronizableDataLocalIdModel<Long> {
+public class TreatmentQuestionDataModel extends SynchronizableDataLocalIdModel<Long> {
     @NotNull(message = ID_NULL_MESSAGE)
     private Long treatmentId;
 
@@ -15,7 +15,7 @@ public class FaqUserQuestionDataModel extends SynchronizableDataLocalIdModel<Lon
     @Size(min = 1, max = TREATMENT_USER_QUESTION_MAX, message = QUESTION_SIZE_MESSAGE)
     private String question;
 
-    public FaqUserQuestionDataModel() {
+    public TreatmentQuestionDataModel() {
     }
 
     public Long getTreatmentId() {
